pyproject.toml
[tool.poetry]
name = "sample-rag"
version = "0.1.0"
description = ""
authors = ["Alexey Nikitin <alexeynktn@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypdf = "^5.3.1"
python-dotenv = "^1.0.1"
openai = "^1.66.3"
pydbantic = "^0.0.47"
aiosqlite = "^0.21.0"
black = "^25.1.0"
isort = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
